refactor: Split functions, tests
